+++ reviews/admin.py
@@ -1,3 +1,4 @@
 from django.contrib import admin
+from .models import Review
 
-# Register your models here.
+admin.site.register(Review)
