+++ reviews/forms/review_form.py
@@ -1,5 +1,6 @@
 from django import forms
 
+
 class ReviewForm(forms.Form):
 
     placeholders = {
